TestScreen wurde ersetzt durch integration in MainActivty
